# backend/src/endpoints/claims.ts
import { Hono } from 'hono'
import { INSURER_ROLES, type AppEnv } from '../types'
import { requireRole } from '../security/rbac'
import { writeAuditEvent } from '../security/audit'
import { loadAuthorizedClaim, transitionClaim } from '../security/claimAccess'
import { MAIN_PATH, isClaimStage } from '../security/claimStateMachine'
import { destinationHash, latestDecision, payoutFor } from '../security/ledger'
import {
  appealSchema,
  claimIdParam,
  initiateClaimSchema,
  listQuerySchema,
  payoutDetailsSchema,
  screeningSchema,
  validate,
  verifyEligibilitySchema,
} from '../security/validation'

// Customer-side claim routes. Insurer-side transitions live in claimsInsurer.ts.
const router = new Hono<AppEnv>()

const notFound = { error: 'not_found' } as const

async function findOwnedPolicy(db: D1Database, policyId: string, userId: string) {
  return db
    .prepare('SELECT id, status, tenant_id FROM policies WHERE id = ? AND user_id = ?')
    .bind(policyId, userId)
    .first<{ id: string; status: string; tenant_id: string | null }>()
}

// General status check
router.get('/status', (c) => c.json({ status: 'active' }))

// List claims visible to the actor: a customer sees their own, insurer staff see their tenant's.
router.get('/', validate('query', listQuerySchema), async (c) => {
  const actor = c.get('actor')
  const { limit } = c.req.valid('query')

  if (actor.role === 'CUSTOMER') {
    const { results } = await c.env.DB.prepare(
      `SELECT id, policy_id, tenant_id, stage, status, category, claimed_amount_cents, created_at, updated_at
       FROM claims WHERE user_id = ? ORDER BY created_at DESC, id LIMIT ?`
    )
      .bind(actor.id, limit)
      .all()
    return c.json({ claims: results })
  }

  if (INSURER_ROLES.includes(actor.role) && actor.tenantId !== null) {
    // Drafts are not yet shared with the insurer. user_id is a platform-wide customer id
    // and is not exposed to tenant staff (DECISION REQUIRED: per-tenant claimant reference).
    const { results } = await c.env.DB.prepare(
      `SELECT id, policy_id, tenant_id, stage, status, category, claimed_amount_cents, created_at, updated_at
       FROM claims WHERE tenant_id = ? AND stage <> 'Draft' ORDER BY created_at DESC, id LIMIT ?`
    )
      .bind(actor.tenantId, limit)
      .all()
    return c.json({ claims: results })
  }

  if (actor.role === 'SUPERADMIN') {
    // Platform read-only view across insurers; Drafts stay private. Optional ?tenantId filter.
    const tenantId = c.req.query('tenantId') ?? null
    const { results } = await c.env.DB.prepare(
      `SELECT id, policy_id, tenant_id, stage, status, category, claimed_amount_cents, created_at, updated_at
       FROM claims WHERE stage <> 'Draft' AND (? IS NULL OR tenant_id = ?) ORDER BY created_at DESC, id LIMIT ?`
    )
      .bind(tenantId, tenantId, limit)
      .all()
    return c.json({ claims: results })
  }

  await writeAuditEvent(c, { action: 'authz.role_denied', resourceType: 'route', resourceId: 'GET /claims', outcome: 'denied' })
  return c.json({ error: 'forbidden' }, 403)
})

// Wizard Step 2: Verification Check
// Only policy ownership + Active status can be verified today. Identity and
// waiting-period checks are not implemented, so they are reported as null (unknown)
// rather than true.
router.post('/verify-eligibility', requireRole('CUSTOMER'), validate('json', verifyEligibilitySchema), async (c) => {
  const { policyId } = c.req.valid('json')
  const policy = await findOwnedPolicy(c.env.DB, policyId, c.get('actor').id)
  const isPolicyActive = policy?.status === 'Active'
  return c.json({
    verified: isPolicyActive,
    context: { isIdentityValid: null, isPolicyActive, waitingPeriodCleared: null },
  })
})

// Wizard Step 1: Initiate (Draft creation)
router.post('/initiate', requireRole('CUSTOMER'), validate('json', initiateClaimSchema), async (c) => {
  const actor = c.get('actor')
  const { policyId, category } = c.req.valid('json')

  const policy = await findOwnedPolicy(c.env.DB, policyId, actor.id)
  const reason = !policy
    ? 'policy_not_owned'
    : policy.status !== 'Active'
      ? 'policy_not_active'
      : policy.tenant_id === null
        ? 'policy_missing_tenant'
        : null
  if (!policy || reason) {
    await writeAuditEvent(c, {
      action: 'claim.initiate_rejected',
      resourceType: 'policy',
      resourceId: policyId,
      outcome: 'denied',
      details: { reason },
    })
    return c.json({ error: 'policy_not_eligible' }, 422)
  }

  // Unguessable ID (previously claim_${Date.now()}, which was enumerable).
  const claimId = `claim_${crypto.randomUUID()}`
  const now = new Date().toISOString()
  // tenant_id is copied from the policy's insurer, never taken from the client.
  await c.env.DB.prepare(
    `INSERT INTO claims (id, user_id, policy_id, tenant_id, stage, status, category, created_at, updated_at)
     VALUES (?, ?, ?, ?, 'Draft', 'Pending', ?, ?, ?)`
  )
    .bind(claimId, actor.id, policyId, policy.tenant_id, category ?? null, now, now)
    .run()

  await writeAuditEvent(c, {
    action: 'claim.created',
    resourceType: 'claim',
    resourceId: claimId,
    outcome: 'success',
    details: { policyId, tenantId: policy.tenant_id },
  })
  return c.json({ status: 'draft_created', claimId }, 201)
})

// Wizard Step 3: Screening & Context
router.patch(
  '/:claimId/screening',
  requireRole('CUSTOMER'),
  validate('param', claimIdParam),
  validate('json', screeningSchema),
  async (c) => {
    const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'owner-write')
    if (!claim) return c.json(notFound, 404)

    if (claim.stage !== 'Draft' && claim.stage !== 'Info Needed') {
      return c.json({ error: 'claim_not_editable', stage: claim.stage }, 409)
    }

    const { causeOfLoss, incidentDate } = c.req.valid('json')
    // The stage precondition is repeated in SQL so a concurrent transition cannot slip in
    // between the read above and this write.
    const written = await c.env.DB.prepare(
      "UPDATE claims SET cause_of_loss = ?, incident_date = ?, updated_at = ? WHERE id = ? AND stage IN ('Draft', 'Info Needed')"
    )
      .bind(causeOfLoss, incidentDate, new Date().toISOString(), claim.id)
      .run()
    if (written.meta.changes !== 1) return c.json({ error: 'claim_not_editable' }, 409)
    await writeAuditEvent(c, {
      action: 'claim.screening_updated',
      resourceType: 'claim',
      resourceId: claim.id,
      outcome: 'success',
    })

    // Answering an information request sends the claim back into screening.
    if (claim.stage === 'Info Needed') {
      const t = await transitionClaim(c, claim, 'Screening')
      if (!t.ok) return c.json({ error: t.error }, t.status)
    }

    return c.json({ status: 'screening_updated', message: 'Cause of loss and context saved.' })
  }
)

// Wizard Step 4: Supporting Evidence (OCR Queue)
// NOTE: no file is accepted or stored yet (no storage binding exists). This only
// queues an event. See docs/security/EVIDENCE_SECURITY.md.
router.post('/:claimId/evidence-ocr', requireRole('CUSTOMER'), validate('param', claimIdParam), async (c) => {
  const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'owner-write')
  if (!claim) return c.json(notFound, 404)

  if (claim.stage !== 'Draft' && claim.stage !== 'Info Needed') {
    return c.json({ error: 'claim_not_editable', stage: claim.stage }, 409)
  }

  await c.env.CLAIM_EVENTS?.send({
    event: 'ClaimEvidenceUploaded',
    data: { claimId: claim.id, timestamp: new Date().toISOString() },
  })
  await writeAuditEvent(c, {
    action: 'claim.evidence_queued',
    resourceType: 'claim',
    resourceId: claim.id,
    outcome: 'success',
  })
  return c.json({ status: 'processing OCR', message: 'Documents queued for analysis' }, 202)
})

// Wizard Step 5: Submit for Review
router.post('/:claimId/submit', requireRole('CUSTOMER'), validate('param', claimIdParam), async (c) => {
  const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'owner-write')
  if (!claim) return c.json(notFound, 404)

  if (claim.stage === 'Draft' && (!claim.cause_of_loss || !claim.incident_date)) {
    return c.json({ error: 'screening_incomplete' }, 422)
  }

  const t = await transitionClaim(c, claim, 'Submitted')
  if (!t.ok) return c.json({ error: t.error }, t.status)

  await c.env.CLAIM_EVENTS?.send({
    event: 'ClaimSubmitted',
    data: { claimId: claim.id, timestamp: new Date().toISOString() },
  })
  return c.json({ status: 'submitted', message: 'Claim successfully submitted for decision.', claimId: claim.id })
})

/**
 * Claim detail for the owning customer or the claim's tenant staff (same 'read' access as
 * /timeline). One object for the app's claim screens. Never returns user_id (platform-wide
 * customer id, BACKEND-SEC-017) or the payout destination hash; the destination is masked.
 */
router.get('/:claimId', validate('param', claimIdParam), async (c) => {
  const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'read')
  if (!claim) return c.json(notFound, 404)
  const context = await c.env.DB.prepare(
    `SELECT p.plan_name AS plan_name, t.name AS insurer_name
     FROM policies p LEFT JOIN tenants t ON t.id = p.tenant_id WHERE p.id = ?`
  )
    .bind(claim.policy_id)
    .first<{ plan_name: string | null; insurer_name: string | null }>()
  return c.json({
    claim: {
      id: claim.id,
      policyId: claim.policy_id,
      planName: context?.plan_name ?? null,
      tenantId: claim.tenant_id,
      insurerName: context?.insurer_name ?? null,
      stage: isClaimStage(claim.stage) ? claim.stage : 'Unknown',
      status: claim.status,
      category: claim.category,
      causeOfLoss: claim.cause_of_loss,
      incidentDate: claim.incident_date,
      claimedAmountCents: claim.claimed_amount_cents,
      payoutDestination: claim.payout_account_last4
        ? { bankName: claim.payout_bank_name, accountLast4: claim.payout_account_last4 }
        : null,
      createdAt: claim.created_at,
      updatedAt: claim.updated_at,
    },
  })
})

// Timeline and Decision (For Step 6: Status & Tracking)
router.get('/:claimId/timeline', validate('param', claimIdParam), async (c) => {
  const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'read')
  if (!claim) return c.json(notFound, 404)

  const { results } = await c.env.DB.prepare(
    `SELECT occurred_at, details FROM audit_events
     WHERE resource_type = 'claim' AND resource_id = ? AND action = 'claim.stage_changed'
     ORDER BY occurred_at`
  )
    .bind(claim.id)
    .all<{ occurred_at: string; details: string }>()

  const reachedAt = new Map<string, string>()
  for (const row of results) {
    const to = (JSON.parse(row.details) as { to?: string }).to
    if (to) reachedAt.set(to, row.occurred_at)
  }

  // A stage counts as completed when the audit trail shows it was reached, or when the claim
  // is currently at or beyond it on the main path (seeded claims have no audit history).
  const currentIndex = MAIN_PATH.indexOf(claim.stage as (typeof MAIN_PATH)[number])
  return c.json({
    claimId: claim.id,
    currentStage: isClaimStage(claim.stage) ? claim.stage : 'Unknown',
    timeline: MAIN_PATH.map((stage, i) => ({
      stage,
      date: reachedAt.get(stage) ?? null,
      completed: reachedAt.has(stage) || (currentIndex >= 0 && i <= currentIndex),
    })),
  })
})

router.get('/:claimId/decision', validate('param', claimIdParam), async (c) => {
  const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'read')
  if (!claim) return c.json(notFound, 404)

  // Phase 3: the decision record is authoritative. Claims decided before Phase 3 (seed data)
  // have no record; their claims.status is reported with record: null.
  const record = await latestDecision(c.env.DB, claim.id)
  if (record) {
    return c.json({
      decision: record.outcome,
      record: {
        id: record.id,
        decidedAt: record.decided_at,
        decidedByRole: record.actor_role,
        reason: record.reason,
        approvedAmountCents: record.approved_amount_cents,
        rulesVersion: record.rules_version,
        evidenceDigest: record.evidence_digest,
      },
    })
  }
  const decided = claim.stage === 'Decision' || claim.stage === 'Paid'
  return c.json({ decision: decided ? claim.status : 'pending', record: null })
})

/**
 * Phase 3: the customer states what they claim and where a payout should go, while the
 * claim is still theirs to edit (Draft / Info Needed). Once submitted these inputs are
 * frozen; the decision snapshots the destination hash and /pay refuses to pay to anything
 * else. The full account number is hashed and never stored or returned.
 */
router.put(
  '/:claimId/payout-details',
  requireRole('CUSTOMER'),
  validate('param', claimIdParam),
  validate('json', payoutDetailsSchema),
  async (c) => {
    const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'owner-write')
    if (!claim) return c.json(notFound, 404)

    if (claim.stage !== 'Draft' && claim.stage !== 'Info Needed') {
      await writeAuditEvent(c, {
        action: 'payout.destination_change_blocked',
        resourceType: 'claim',
        resourceId: claim.id,
        outcome: 'denied',
        details: { stage: claim.stage, reason: 'locked_after_submission' },
      })
      return c.json({ error: 'payout_details_locked', stage: claim.stage }, 409)
    }

    const { claimedAmountCents, bankName, accountHolder, accountNumber } = c.req.valid('json')
    const hash = await destinationHash(bankName, accountNumber)
    const last4 = accountNumber.slice(-4)
    const written = await c.env.DB.prepare(
      `UPDATE claims SET claimed_amount_cents = ?, payout_bank_name = ?, payout_account_holder = ?,
         payout_account_last4 = ?, payout_destination_hash = ?, payout_details_updated_at = ?, updated_at = ?
       WHERE id = ? AND stage IN ('Draft', 'Info Needed')`
    )
      .bind(claimedAmountCents, bankName.trim(), accountHolder.trim(), last4, hash, new Date().toISOString(), new Date().toISOString(), claim.id)
      .run()
    if (written.meta.changes !== 1) return c.json({ error: 'payout_details_locked' }, 409)

    await writeAuditEvent(c, {
      action: 'claim.payout_details_set',
      resourceType: 'claim',
      resourceId: claim.id,
      outcome: 'success',
      details: { claimedAmountCents, destinationLast4: last4 },
    })
    return c.json({ status: 'payout_details_saved', claimedAmountCents, destination: { bankName: bankName.trim(), accountLast4: last4 } })
  }
)

/** Masked view of the money side of a claim: inputs, latest decision, payout (if any). */
router.get('/:claimId/payout', validate('param', claimIdParam), async (c) => {
  const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'read')
  if (!claim) return c.json(notFound, 404)

  const [decision, payout] = await Promise.all([latestDecision(c.env.DB, claim.id), payoutFor(c.env.DB, claim.id)])
  return c.json({
    claimId: claim.id,
    stage: claim.stage,
    claimedAmountCents: claim.claimed_amount_cents,
    destination: claim.payout_destination_hash
      ? { bankName: claim.payout_bank_name, accountHolder: claim.payout_account_holder, accountLast4: claim.payout_account_last4 }
      : null,
    decision: decision
      ? { id: decision.id, outcome: decision.outcome, approvedAmountCents: decision.approved_amount_cents, decidedAt: decision.decided_at, decidedByRole: decision.actor_role }
      : null,
    payout: payout
      ? { id: payout.id, amountCents: payout.amount_cents, destinationLast4: payout.destination_last4, status: payout.status, initiatedAt: payout.initiated_at }
      : null,
  })
})

router.post(
  '/:claimId/appeal',
  requireRole('CUSTOMER'),
  validate('param', claimIdParam),
  validate('json', appealSchema),
  async (c) => {
    const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'owner-write')
    if (!claim) return c.json(notFound, 404)

    // Only a rejected decision can be appealed.
    if (claim.stage !== 'Decision' || claim.status !== 'Rejected') {
      return c.json({ error: 'not_appealable' }, 409)
    }

    const t = await transitionClaim(c, claim, 'Appeal')
    if (!t.ok) return c.json({ error: t.error }, t.status)
    return c.json({ status: 'appealed' })
  }
)

router.post(
  '/:claimId/withdraw',
  requireRole('CUSTOMER'),
  validate('param', claimIdParam),
  async (c) => {
    const claim = await loadAuthorizedClaim(c, c.req.valid('param').claimId, 'owner-write')
    if (!claim) return c.json(notFound, 404)

    const t = await transitionClaim(c, claim, 'Withdrawn')
    if (!t.ok) return c.json({ error: t.error }, t.status)
    return c.json({ status: 'withdrawn' })
  }
)

export default router